+++ checker-qual/src/main/java/org/checkerframework/framework/qual/PurityUnqualified.java
@@ -1,4 +1,4 @@
-package org.checkerframework.framework.util;
+package org.checkerframework.framework.qual;
 
 import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
