+++ framework/src/main/java/org/checkerframework/framework/util/PurityAnnotatedTypeFactory.java
@@ -6,7 +6,9 @@ import java.util.HashSet;
 import java.util.Set;
 import org.checkerframework.common.basetype.BaseAnnotatedTypeFactory;
 import org.checkerframework.common.basetype.BaseTypeChecker;
+import org.checkerframework.framework.qual.PurityUnqualified;
 
+/** AnnotatedTypeFactory for the {@link PurityChecker}. */
 public class PurityAnnotatedTypeFactory extends BaseAnnotatedTypeFactory {
 
     public PurityAnnotatedTypeFactory(BaseTypeChecker checker) {
